refactor: changed login to adhere to page component
