+++ frontend/src/pages/Login/LoginPage.tsx
@@ -2,7 +2,7 @@ import React, {Dispatch, SetStateAction} from 'react';
 import LoginBody from './LoginBody/LoginBody';
 import LoginHeader from './LoginHeader';
 import './Login.css'
-import { LoginProp } from '../data/LoginProp'
+import { LoginProp } from '../../data/LoginProp'
 
 export default function Login({ setToken }: LoginProp) {
     return (
